[data-table] changed how to calculate total rows and generate UNIQ keys
