+++ semcore/data-table/src/components/Body/Body.tsx
@@ -9,12 +9,11 @@ import { Cell } from './Cell';
 import { DataTableRowProps, DTRow, RowPropsInner } from './Row.types';
 import { DataTableCellProps } from './Cell.types';
 import { MergedColumnsCell, MergedRowsCell } from './MergedCells';
-import { ACCORDION, ROW_GROUP, UNIQ_ROW_KEY } from '../DataTable/DataTable';
+import { ACCORDION, UNIQ_ROW_KEY } from '../DataTable/DataTable';
 import ChevronRightM from '@semcore/icon/ChevronRight/m';
 import { ButtonLink } from '@semcore/button';
-import { DataRowItem, DataTableData, DTValue } from '../DataTable/DataTable.types';
+import { DataTableData, DTValue } from '../DataTable/DataTable.types';
 import Spin from '@semcore/spin';
-import { DTColumn } from '../Head/Column.types';
 
 const ROWS_BUFFER = 20;
 const APROX_ROWS_ON_PAGE = 20;
@@ -24,32 +23,15 @@ class BodyRoot<D extends DataTableData> extends Component<
   {},
   {},
   [],
-  BodyPropsInner<D>
+  BodyPropsInner
 > {
   static displayName = 'Body';
   static style = style;
-
-  private columnsSplitter = '/';
-  private rows: Array<DTRow | DTRow[]> = [];
 
   rowsHeightMap = new Map<number, [number, number, HTMLElement]>();
 
   indexForDownIterate = 0;
   indexForUpIterate = 0;
-
-  componentDidMount() {
-    this.rows = this.calculateRows();
-
-    this.forceUpdate();
-  }
-
-  componentDidUpdate(prevProps: DataTableBodyProps & BodyPropsInner<D>) {
-    if (prevProps.data !== this.asProps.data) {
-      this.rows = this.calculateRows();
-
-      this.forceUpdate();
-    }
-  }
 
   handleRef = (index: number, row: DTRow) => (node: HTMLElement | null) => {
     if (!this.rowsHeightMap.has(index) && node) {
@@ -83,10 +65,11 @@ class BodyRoot<D extends DataTableData> extends Component<
       uid,
       onBackFromAccordion,
       rowProps,
+      rows,
+      flatRows,
     } = this.asProps;
     const row = props.row;
     const index = props.offset + i;
-    const flatRows = this.rows.flat();
 
     const rowIndex = Array.from(expandedRows ?? []).reduce((acc, item) => {
       const rowIndex = flatRows.findIndex((row) => row[UNIQ_ROW_KEY] === item);
@@ -122,7 +105,7 @@ class BodyRoot<D extends DataTableData> extends Component<
       rowIndex: index,
       ariaRowIndex,
       gridRowIndex,
-      rows: this.rows,
+      rows,
       onBackFromAccordion,
       row,
       expandedRows,
@@ -225,9 +208,10 @@ class BodyRoot<D extends DataTableData> extends Component<
       scrollDirection,
       tableContainerRef,
       scrollTop,
+      rows,
     } = this.asProps;
 
-    let rowsToRender = this.rows;
+    let rowsToRender = rows;
     let startIndex = -1;
     let lastIndex = -1;
 
@@ -258,7 +242,7 @@ class BodyRoot<D extends DataTableData> extends Component<
             }
 
             if (startIndex !== -1 && scrollTop + offsetHeight < valueToToCompare) {
-              lastIndex = Math.min(key + nextPrepared, this.rows.length);
+              lastIndex = Math.min(key + nextPrepared, rows.length);
             }
 
             if (startIndex !== -1 && lastIndex !== -1) {
@@ -278,7 +262,7 @@ class BodyRoot<D extends DataTableData> extends Component<
             const valueToToCompare = value[0];
 
             if (lastIndex === -1 && scrollTop + offsetHeight > valueToToCompare) {
-              lastIndex = Math.min(key + nextPrepared, this.rows.length);
+              lastIndex = Math.min(key + nextPrepared, rows.length);
             }
 
             if (lastIndex !== -1 && scrollTop < valueFromToCompare) {
@@ -296,17 +280,17 @@ class BodyRoot<D extends DataTableData> extends Component<
         }
 
         if (startIndex === -1) {
-          startIndex = scrollTop === 0 ? 0 : Math.max(this.rows.length - aproxRowsOnPage, 0);
+          startIndex = scrollTop === 0 ? 0 : Math.max(rows.length - aproxRowsOnPage, 0);
         }
 
         if (lastIndex === -1) {
-          lastIndex = scrollTop === 0 ? aproxRowsOnPage : this.rows.length;
+          lastIndex = scrollTop === 0 ? aproxRowsOnPage : rows.length;
         }
 
         this.indexForDownIterate = startIndex;
         this.indexForUpIterate = lastIndex;
 
-        rowsToRender = this.rows.slice(startIndex, lastIndex);
+        rowsToRender = rows.slice(startIndex, lastIndex);
       } else if ('rowHeight' in virtualScroll) {
         const rowHeight = virtualScroll.rowHeight;
 
@@ -314,10 +298,10 @@ class BodyRoot<D extends DataTableData> extends Component<
 
         const lastIndex = Math.min(
           Math.ceil((scrollTop + offsetHeight) / rowHeight) + nextPrepared,
-          this.rows.length,
+          rows.length,
         );
 
-        rowsToRender = this.rows.slice(startIndex, lastIndex);
+        rowsToRender = rows.slice(startIndex, lastIndex);
       }
     }
 
@@ -376,108 +360,6 @@ class BodyRoot<D extends DataTableData> extends Component<
       obj === undefined ||
       obj === null
     );
-  }
-
-  private calculateRows(): Array<DTRow[] | DTRow> {
-    const { data, uid, columns } = this.asProps;
-
-    const rows: Array<DTRow[] | DTRow> = [];
-    const columnNames = columns.map((column: DTColumn) => column.name);
-
-    let rowIndex = 0;
-
-    const id = 100000000; // need this for gen keys by toString(36)
-
-    const makeDtRow = (row: DataRowItem, excludeColumns?: string[]) => {
-      const columns = new Set(columnNames);
-      const dtRow = Object.entries(row).reduce<DTRow>(
-        (acc, [key, value]) => {
-          const columnsToRow = key.split(this.columnsSplitter);
-
-          if (columnsToRow.length === 1) {
-            acc[key] = value ?? '';
-            columns.delete(key);
-          } else {
-            acc[columnsToRow[0]] = new MergedColumnsCell(value, {
-              dataKey: key,
-              size: columnsToRow.length,
-            });
-            columnsToRow.forEach((value) => {
-              columns.delete(value);
-            });
-          }
-
-          if (row[ACCORDION]) {
-            acc[ACCORDION] = row[ACCORDION];
-          }
-
-          return acc;
-        },
-        {
-          [UNIQ_ROW_KEY]: row[UNIQ_ROW_KEY] ?? `${uid}_${(rowIndex + id).toString(36)}`,
-        },
-      );
-
-      excludeColumns?.forEach((value) => {
-        columns.delete(value);
-      });
-
-      if (columns.size > 0) {
-        columns.forEach((value) => {
-          dtRow[value] = '';
-        });
-      }
-
-      return dtRow;
-    };
-
-    data.forEach((row) => {
-      const groupedRows: DataTableData | undefined = row[ROW_GROUP];
-
-      const fromRow = rowIndex + 2; // 1 - for header, 1 - because start not from 0, but from 1
-
-      if (groupedRows) {
-        const toRow = fromRow + groupedRows.length;
-        const innerRows: DTRow[] = [];
-
-        const groupedKeys: string[] = [];
-        const groupedRowData = Object.entries(row).reduce<DTRow>(
-          (acc, [key, value]) => {
-            acc[key] = new MergedRowsCell(value, [fromRow, toRow]);
-            groupedKeys.push(key);
-            return acc;
-          },
-          {
-            [UNIQ_ROW_KEY]: '', // will fill in makeDtRow
-          },
-        );
-
-        groupedRows.forEach((childRow, index) => {
-          let dtRow: DTRow;
-          if (index === 0) {
-            const rowData = {
-              ...childRow,
-              ...groupedRowData,
-            };
-            dtRow = makeDtRow(rowData);
-          } else {
-            dtRow = makeDtRow(childRow, groupedKeys);
-          }
-
-          innerRows.push(dtRow);
-          rowIndex++;
-        });
-
-        rows.push(innerRows);
-      } else {
-        const dtRow = makeDtRow(row);
-
-        rows.push(dtRow);
-        rowIndex++;
-      }
-    });
-
-    return rows;
   }
 
   private setRowHeight(index: number, row: DTRow) {
